Move alpha app gateways to Alpha folder
