+++ src/app/Gateways/Alphas/CreateDocumentRequest.test.js
@@ -1,7 +1,7 @@
 import createDocumentRequest from './CreateDocumentRequest';
 import { enableFetchMocks } from 'jest-fetch-mock';
-import { hackneyToken } from '../lib/Cookie';
-jest.mock('../lib/Cookie');
+import { hackneyToken } from '../../lib/Cookie';
+jest.mock('../../lib/Cookie');
 
 describe('CreateDocumentRequest', () => {
   beforeEach(() => {
