+++ src/app/Gateways/Alphas/CreateVulnerability.test.js
@@ -1,7 +1,7 @@
 import createVulnerability from './CreateVulnerability';
 import { enableFetchMocks } from 'jest-fetch-mock';
-import { hackneyToken } from '../lib/Cookie';
-jest.mock('../lib/Cookie');
+import { hackneyToken } from '../../lib/Cookie';
+jest.mock('../../lib/Cookie');
 
 describe('CreateVulnerability', () => {
   beforeEach(() => {
